+++ src/CryptoTracker/CryptoTracker/Import/Objects/BitcoinDeTransaction.cs
@@ -37,7 +37,7 @@ namespace CryptoTracker.Import.Objects
         /// <summary>
         /// Kurs in EUR (z.B. 61 EUR pro BTC)
         /// </summary>
-        public double? Kurs { get; set; }
+        public decimal? Kurs { get; set; }
 
         /// <summary>
         /// z.B. BTC / EUR
@@ -49,13 +49,13 @@ namespace CryptoTracker.Import.Objects
         /// z.B. Anzahl an BTC/ETH vor Gebührenabzug. z.B. 1.000
         /// </summary>
         [Name("Crypto vor Gebühr")]
-        public double? CryptoVorGebuehr { get; set; }
+        public decimal? CryptoVorGebuehr { get; set; }
 
         /// <summary>
         /// z.B. Anzahl der Gegenwährung (z.B. EUR) vor Gebührenabzug z.B. 61€
         /// </summary>
         [Name("Menge vor Gebühr")]
-        public double? MengeVorGebuehr { get; set; }
+        public decimal? MengeVorGebuehr { get; set; }
 
         /// <summary>
         /// z.B. EUR
@@ -67,13 +67,13 @@ namespace CryptoTracker.Import.Objects
         /// z.B. Anzahl an BTC/ETH nach Gebührenabzug. z.B. 1.000
         /// </summary>
         [Name("Crypto nach Bitcoin.de-Gebühr")]
-        public double? CryptoNachGebuehr { get; set; }
+        public decimal? CryptoNachGebuehr { get; set; }
 
         /// <summary>
         /// z.B. Anzahl der Gegenwährung (z.B. EUR) nach Gebührenabzug z.B. 61€
         /// </summary>
         [Name("Menge nach Bitcoin.de-Gebühr")]
-        public double? MengeNachGebuehr { get; set; }
+        public decimal? MengeNachGebuehr { get; set; }
 
         /// <summary>
         /// z.B. EUR
@@ -85,12 +85,12 @@ namespace CryptoTracker.Import.Objects
         /// Menge, die vom Konto zu- oder abgebucht wurden.
         /// </summary>
         [Name("Zu- / Abgang")]
-        public double ZuAbgang { get; set; }
+        public decimal ZuAbgang { get; set; }
 
         /// <summary>
         /// Kontostand nach dieser Transaktion
         /// </summary>
-        public double Kontostand {  get; set; }
+        public decimal Kontostand {  get; set; }
 
         public string Kommentar { get; set; }
     }
